# api/openapi.yaml
openapi: 3.0.3
info:
  title: Smart Grocery & Meal Planner API
  description: |
    MVP API for managing pantry inventory, grocery lists, meal ideas, and expiration notifications.
    
    ## Features
    - **Inventory Management**: CRUD operations for pantry items with low-stock and expiring-soon flags
    - **Grocery Lists**: Create and manage shopping lists, auto-add low stock or expiring items
    - **Meal Ideas**: Get suggestions based on pantry contents, save favorite meals
    - **Notifications**: Schedule reminders for expiring items
  version: 1.0.0
  contact:
    name: Smart Grocery Team
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/users:
    get:
      tags: [Users]
      summary: List all users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
    post:
      tags: [Users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email already exists
  /api/users/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Users]
      summary: Get user by ID
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Users]
      summary: Update user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUser'
      responses:
        '200':
          description: User updated
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Users]
      summary: Delete user
      responses:
        '200':
          description: User deleted
        '404':
          $ref: '#/components/responses/NotFound'

  # Pantry
...
components:
  schemas:
...
    CreateUser:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          maxLength: 100
        timezone:
          type: string
          maxLength: 50
        dietaryTags:
          type: string
          maxLength: 255
        reminderWindowDays:
          type: integer
          minimum: 1
          maximum: 30
    UpdateUser:
      type: object
      properties:
        name:
          type: string
        timezone:
          type: string
        dietaryTags:
          type: string
        reminderWindowDays:
          type: integer
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/User'
    UserListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        userId:
          type: integer
        email:
          type: string
        name:
          type: string
        timezone:
          type: string
        dietaryTags:
          type: string
        reminderWindowDays:
          type: integer
        createdAt:
          type: string
          format: date-time

    # Pantry schemas
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  message:
                    type: string
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: object
                properties:
                  message:
                    type: string
                  details:
                    type: array
                    items:
                      type: object